chore:  ProblemImageData를 별도의 Model 파일로 분리
